+++ administration/src/mixins/mixinSpellcheck.js
