+++ administration/src/mixins/mixinText.js
